requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.11.7
numpy<2
sentence-transformers==3.0.1
pypdf==4.3.1
python-docx==1.1.2
faiss-cpu==1.8.0
